Add document search

+++ Papyrus.Api.Contracts/Contracts/Api/Pagination.cs
@@ -2,7 +2,7 @@
 
 public record Pagination
 {
-    public Pagination(int pageNumber, int pageSize, long total)
+    public Pagination(int pageNumber, int pageSize, int total)
     {
         Page = pageNumber;
         Size = pageSize;
@@ -11,5 +11,5 @@ public record Pagination
     
     public int Page { get; init; }
     public int Size { get; init; }
-    public long Total { get; init; } 
+    public int Total { get; init; } 
 }
