+++ Papyrus.Persistence.Interfaces/Contracts/PagedResponse.cs
@@ -4,6 +4,6 @@ public record PagedResponse<T>
 {
     public List<T> Data { get; init; } = [];
     
-    public long TotalPages { get; init; }
+    public int TotalPages { get; init; }
     
 }
